gpu: nvgpu: clk: use consistent type for regime id
The clk module was using u8's and u32's for the regime ID. Since the
regime id is only a byte, just use u8's.

This eliminates MISRA rule 10.3 violations for implicit assignments to
different types.

JIRA NVGPU-1008

Reviewed-on: https://git-master.nvidia.com/r/1946259
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/clk/clk.c
@@ -507,7 +507,7 @@ done:
 	return status;
 }
 
-static u32 find_regime_id(struct gk20a *g, u32 domain, u16 clkmhz)
+static u8 find_regime_id(struct gk20a *g, u32 domain, u16 clkmhz)
 {
 	struct fll_device *pflldev;
 	u8 j;
@@ -527,7 +527,7 @@ static u32 find_regime_id(struct gk20a *g, u32 domain, u16 clkmhz)
 	return CTRL_CLK_FLL_REGIME_ID_INVALID;
 }
 
-static int set_regime_id(struct gk20a *g, u32 domain, u32 regimeid)
+static int set_regime_id(struct gk20a *g, u32 domain, u8 regimeid)
 {
 	struct fll_device *pflldev;
 	u8 j;
@@ -543,7 +543,7 @@ static int set_regime_id(struct gk20a *g, u32 domain, u32 regimeid)
 	return -EINVAL;
 }
 
-static int get_regime_id(struct gk20a *g, u32 domain, u32 *regimeid)
+static int get_regime_id(struct gk20a *g, u32 domain, u8 *regimeid)
 {
 	struct fll_device *pflldev;
 	u8 j;
